feat(api): refactor changes
Ref. metriport/metriport-internal#380
