+++ api/app/src/external/commonwell/organization.ts
@@ -2,10 +2,8 @@ import { Organization as CWOrganization } from "@metriport/commonwell-sdk";
 import { Organization } from "../../models/medical/organization";
 import { Config, getEnvVarOrFail } from "../../shared/config";
 import { OID_PREFIX } from "../../shared/oid";
-import { certificate, metriportQueryMeta, makeCommonWelMemberAPI } from "./api";
+import { certificate, metriportQueryMeta, makeCommonWellAPI } from "./api";
 
-// TODO move these "getEnvVarOrFail" to Config
-const metriportOrgName = getEnvVarOrFail("CW_MEMBER_NAME");
 const technicalContact = {
   name: getEnvVarOrFail("CW_TECHNICAL_CONTACT_NAME"),
   title: getEnvVarOrFail("CW_TECHNICAL_CONTACT_TITLE"),
@@ -40,7 +38,7 @@ export async function organizationToCommonwell(
     homeCommunityId: cwId,
     patientIdAssignAuthority: cwId,
     displayName: org.data.name,
-    memberName: metriportOrgName,
+    memberName: Config.getMetriportOrgName(),
     securityTokenKeyType: "BearerKey",
     isActive: true,
     gateways: [
@@ -64,7 +62,10 @@ export async function organizationToCommonwell(
 export const create = async (org: Organization): Promise<void> => {
   const cwOrg = await organizationToCommonwell(org);
   try {
-    const commonWell = makeCommonWelMemberAPI();
+    const commonWell = makeCommonWellAPI(
+      Config.getMemberManagementOID(),
+      Config.getMetriportOrgName()
+    );
     await commonWell.createOrg(metriportQueryMeta, cwOrg);
     await commonWell.addCertificateToOrg(metriportQueryMeta, certificate, org.id);
   } catch (error) {
@@ -78,7 +79,10 @@ export const create = async (org: Organization): Promise<void> => {
 export const update = async (org: Organization): Promise<void> => {
   const cwOrg = await organizationToCommonwell(org);
   try {
-    const commonWell = makeCommonWelMemberAPI();
+    const commonWell = makeCommonWellAPI(
+      Config.getMemberManagementOID(),
+      Config.getMetriportOrgName()
+    );
     await commonWell.updateOrg(metriportQueryMeta, cwOrg, cwOrg.organizationId);
   } catch (error) {
     const msg = `Failure updating Org`;
